requirements.txt
# ─────────────────────────────────────────────────────────────
# 🇮🇳 Indian Trade Intelligence Engine — Python dependencies
# Install: pip install -r requirements.txt
# Python 3.11+ recommended
# ─────────────────────────────────────────────────────────────

# ── Core framework ──────────────────────────────────────────
streamlit>=1.35.0

# ── Supabase (auth + database) ──────────────────────────────
supabase>=2.4.0
psycopg2-binary>=2.9.9          # direct PostgreSQL for dataset upload

# ── AI / LLM ────────────────────────────────────────────────
requests>=2.31.0                # NVIDIA API calls (Llama 3.3 70B)

# ── Data & visualisation ────────────────────────────────────
pandas>=2.1.0
plotly>=5.18.0
numpy>=1.26.0

# ── Excel export ────────────────────────────────────────────
openpyxl>=3.1.2
xlrd>=2.0.1                     # read .xls ITC TradeMap files

# ── PDF export ──────────────────────────────────────────────
reportlab>=4.1.0

# ── HTTP / async ────────────────────────────────────────────
httpx>=0.27.0

# ── Environment / config ────────────────────────────────────
python-dotenv>=1.0.0

# ── Standard library (no install needed) ────────────────────
# hashlib, secrets, smtplib, email, io, os, re, json, datetime, pathlib